fix: fail to parse the 3d model while building
This fixes the TypeError: Failed to parse URL from "FiveFretTrack.fbx".

The reason is Next.js usually tries to render everything on the server (apparently even with the "use client" on the top sometimes), and it lead to an error while trying to fetch the model.

For the fix, I used the next/dynamic, which disables the Server-side Render depending the children of the wrapper. Now Next.js won't try to fetch the model while building.